+++ paymentsheet/src/main/java/com/stripe/android/checkout/BillingDetailsCollectionConfigurationMapper.kt
@@ -1,6 +1,6 @@
 package com.stripe.android.checkout
 
-import com.stripe.android.checkout.PaymentElement.Configuration.BillingDetailsCollectionConfiguration
+import com.stripe.android.elements.PaymentElement.Configuration.BillingDetailsCollectionConfiguration
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import com.stripe.android.paymentsheet.PaymentSheet
