+++ paymentsheet/src/main/java/com/stripe/android/checkout/PaymentMethodLayoutMapper.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.checkout
 
+import com.stripe.android.elements.PaymentElement
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import com.stripe.android.paymentsheet.PaymentSheet
